Interpret keyframe interval value as milliseconds.
Change storage type from double to int.

+++ http_client/video_encoder.h
@@ -109,7 +109,7 @@ class VideoFrameCallbackInterface {
 
 struct VpxConfig {
   // Time between keyframes, in seconds.
-  double keyframe_interval;
+  int keyframe_interval;
   // Video bitrate, in kilobits.
   int bitrate;
   // Video frame rate decimation factor.
